% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weighted-ridge.R
\name{predictWtRidge}
\alias{predictWtRidge}
\title{Predict Weighted Ridge Linear Model}
\usage{
predictWtRidge(model, x)
}
\arguments{
\item{model}{object of class \code{wt_ridge}.}

\item{x}{matrix containing new data instances.}
}
\value{
class labels and membership probabilities.
}
\description{
Predict response for new data instances.
}

