% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weighted-ridge.R
\name{weightedRidgeLinMod}
\alias{weightedRidgeLinMod}
\title{Weighted Ridge Linear Model}
\usage{
weightedRidgeLinMod(x, y, w, lambda, balance = TRUE)
}
\arguments{
\item{x}{design matrix.}

\item{y}{\code{factor} with class labels.}

\item{w}{sample weights (default: \code{rep(1/N, N)}).}

\item{lambda}{optional. L2 penalty parameter.}

\item{balance}{use the inverse of the class frequencies as weights (default: \code{TRUE}).}
}
\value{
object of class \code{wt_ridge}.
}
\description{
Fit weighted ridge regression on inbalanced data.
}

